Refine Druid auto-configuration tests
Split the Spring Boot and Spring Cloud auto-configuration coverage into focused condition tests and full integration tests. This adds microsphere-spring-boot-test for the new AutoConfigurationTest-based cases, verifies disabled and missing-class scenarios, and keeps end-to-end assertions for Druid beans, filters, metadata, and Cloud features. It also normalizes the Druid availability condition annotation declaration.
