+++ microsphere-alibaba-druid-spring-boot/pom.xml
@@ -89,6 +89,13 @@
             <scope>test</scope>
         </dependency>
 
+        <!-- Microsphere Spring Boot Test -->
+        <dependency>
+            <groupId>io.github.microsphere-projects</groupId>
+            <artifactId>microsphere-spring-boot-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+
         <!-- Microsphere Alibaba Druid Spring Test -->
         <dependency>
             <groupId>io.github.microsphere-projects</groupId>
