+++ microsphere-alibaba-druid-spring-cloud/pom.xml
@@ -96,6 +96,13 @@
             <scope>test</scope>
         </dependency>
 
+        <!-- Microsphere Spring Boot Test -->
+        <dependency>
+            <groupId>io.github.microsphere-projects</groupId>
+            <artifactId>microsphere-spring-boot-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+
         <!-- Microsphere Alibaba Druid Spring Test -->
         <dependency>
             <groupId>io.github.microsphere-projects</groupId>
